fix: address third round of Copilot review comments
- messages_wasm.go: use exitcodes.TypeName(code) in Marshal-error
  fallback so the type field is always correct, not hard-coded
- megaport_wasm.go: read --output from parsed flags (not global state)
  so JSON-mode detection is reliable even when action setup hasn't run;
  remove now-unused output import
- output_test.go/messages_native_test.go: move PrintErrorJSON tests into
  a !wasm-tagged file so they compile correctly under -tags js,wasm and
  use the captureStderr helper that is already guarded by that tag
- output_test.go: restore original format (not hard-coded "table") in
  TestGetOutputFormat cleanup via t.Cleanup
- utils_test.go: fix captureStderr named-return so result is set after
  the drain goroutine finishes; cleanup (w.Close, <-done, r.Close) now
  all happen in a single defer so runtime.Goexit cannot leave the pipe open
